fix kill menu notifs

+++ desktop/.config/kill_menu.fish
@@ -1,5 +1,8 @@
 #!/usr/bin/env fish
 
 set -l program $(ps -eo comm= | fsel --dmenu)
-pkill $program
-notify-send "killed $program"
+
+if test -n "$program"
+    pkill $program
+    notify-send "killed $program"
+end
